# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "description": "Create a new user with the given information. Only admin can create users",
        "operationId": "UsersController_createUser_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Returns a paginated list of users. Admins may filter by type and by a case-insensitive search on name or userName.",
        "operationId": "UsersController_listUsers_v1",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Return only users of this type.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "provider",
                "admin"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on display name or login name (userName).",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 50, max 100).",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip (pagination).",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "description": "Update a user with the given information",
        "operationId": "UsersController_updateUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Get a user with the given ID",
        "operationId": "UsersController_getUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "includeAccounts",
            "required": false,
            "in": "query",
            "description": "When true, includes each account id, balance, lock state, public key, and NFC UID (no secrets).",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "description": "Get the current user",
        "operationId": "UsersController_getCurrentUser_v1",
        "parameters": [
          {
            "name": "includeAccounts",
            "required": false,
            "in": "query",
            "description": "When true, includes each account id, balance, lock state, public key, and NFC UID (no secrets).",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Classurency API",
    "description": "API documentation for the Classurency backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "token",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace",
            "maxLength": 128,
            "description": "User display name."
          },
          "userName": {
            "type": "string",
            "description": "User username (letters, digits, - _ + only; no spaces)",
            "maxLength": 128,
            "pattern": "^[-a-zA-Z0-9_+]+$"
          },
          "type": {
            "type": "string",
            "enum": [
              "user",
              "provider",
              "admin"
            ],
            "description": "User role. Defaults to user when omitted."
          }
        },
        "required": [
          "name",
          "userName"
        ]
      },
      "CreateUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User name"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "type": {
            "type": "string",
            "description": "User type",
            "enum": [
              "user",
              "provider",
              "admin"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "password",
          "type"
        ]
      },
      "ListUserItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "user",
              "provider",
              "admin"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "userName",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListUsersResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListUserItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters (ignoring limit/offset)."
          }
        },
        "required": [
          "users",
          "total"
        ]
      },
      "UpdateUserRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace",
            "maxLength": 128,
            "description": "User display name."
          },
          "password": {
            "type": "string",
            "example": "a-long-random-secret",
            "minLength": 6,
            "maxLength": 128,
            "description": "Account password for API login (stored with Argon2id; distinct from the generated card PIN)."
          },
          "type": {
            "type": "string",
            "example": "admin",
            "enum": [
              "user",
              "provider",
              "admin"
            ],
            "description": "User type."
          }
        },
        "required": [
          "name",
          "password",
          "type"
        ]
      },
      "UpdateUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User name"
          },
          "type": {
            "type": "string",
            "description": "User type"
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ]
      },
      "UserAccountSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nfcCardUid": {
            "type": "object",
            "nullable": true,
            "description": "NFC card UID when the account is bound to a card."
          },
          "publicKeyHex": {
            "type": "string",
            "description": "Account public key (hex)."
          },
          "balance": {
            "type": "number",
            "description": "Current balance."
          },
          "isLocked": {
            "type": "boolean",
            "description": "Whether the account is locked."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nfcCardUid",
          "publicKeyHex",
          "balance",
          "isLocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "GetUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User name"
          },
          "type": {
            "type": "string",
            "description": "User type"
          },
          "providerId": {
            "type": "string",
            "format": "uuid",
            "description": "Item provider id when the user type is provider and an item provider exists."
          },
          "accounts": {
            "description": "Included when the request uses includeAccounts=true.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAccountSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ]
      },
...
    }
  }
}