# src/modules/users/users.controller.ts
import {
  Body,
  Controller,
  DefaultValuePipe,
  Get,
  Param,
  ParseBoolPipe,
  Post,
  Patch,
  Query,
  Req,
  UnauthorizedException,
  UseGuards,
  UseInterceptors,
} from '@nestjs/common';
import { ApiBearerAuth, ApiOperation, ApiParam, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
import { RequirePrivilege } from '@common/guards/require-privilege.decorator';
import { ApiTokenGuard } from '@common/guards/api-token.guard';
import { UsersService } from './users.service';
import { CreateUserDto, CreateUserResponseDto } from './dto/create-user.dto';
import { UpdateUserRequestDto, UpdateUserResponseDto } from './dto/update-user.dto';
import { ApiAuthContext } from '@common/guards/api-token.guard';
import { ApiTokenPrivilege } from '@modules/api-tokens/api-token.entity';
import { GetUserRequestDto, GetUserResponseDto } from './dto/get-user.dto';
import { ListUsersQueryDto, ListUsersResponseDto } from './dto/list-users.dto';
import { ResponseDtoOmitter } from '@common/decorators/response-dto-omitter';
import type { Request } from 'express';

@Controller({ path: 'users', version: '1' })
@ApiTags('Users')
@ApiBearerAuth('bearer')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @Post()
  @UseGuards(ApiTokenGuard)
  @RequirePrivilege(ApiTokenPrivilege.ADMIN)
  @ApiOperation({
    summary: 'Create a new user',
    description: 'Create a new user with the given information. Only admin can create users',
  })
  @ApiResponse({ status: 201, description: 'User created successfully', type: CreateUserResponseDto })
  async createUser(@Body() dto: CreateUserDto): Promise<CreateUserResponseDto> {
    return await this.usersService.createUserAsAdmin(dto);
  }

  @Get()
  @UseGuards(ApiTokenGuard)
  @RequirePrivilege(ApiTokenPrivilege.ADMIN)
  @ApiOperation({
    summary: 'List users',
    description:
      'Returns a paginated list of users. Admins may filter by type and by a case-insensitive search on name or userName.',
  })
  @ApiResponse({ status: 200, description: 'Users retrieved successfully', type: ListUsersResponseDto })
  async listUsers(@Query() query: ListUsersQueryDto): Promise<ListUsersResponseDto> {
    return await this.usersService.listUsersAsAdmin(query);
  }

  @Patch(':id')
  @UseGuards(ApiTokenGuard)
  @RequirePrivilege(ApiTokenPrivilege.USER)
  @ApiParam({ name: 'id', description: 'User ID', format: 'uuid' })
  @ApiOperation({
    summary: 'Update a user',
    description: 'Update a user with the given information',
  })
  @ApiResponse({ status: 200, description: 'User updated successfully', type: UpdateUserResponseDto })
  async updateUser(
    @Param('id') userId: string,
    @Body() dto: UpdateUserRequestDto,
    @Req() req: Request & { apiAuth?: ApiAuthContext },
  ): Promise<UpdateUserResponseDto> {
    const reqUserId = req.apiAuth?.userId;
    if (!reqUserId) {
      throw new UnauthorizedException('Missing auth context');
    }
    const isAdmin = req.apiAuth?.privilege === ApiTokenPrivilege.ADMIN;

    return await this.usersService.updateUser(reqUserId, userId, dto, { isAdmin });
  }

  @UseInterceptors(new ResponseDtoOmitter(GetUserResponseDto))
  @Get('me')
  @UseGuards(ApiTokenGuard)
  @RequirePrivilege(ApiTokenPrivilege.USER)
  @ApiOperation({
    summary: 'Get the current user',
    description: 'Get the current user',
  })
  @ApiQuery({
    name: 'includeAccounts',
    required: false,
    type: Boolean,
    description: 'When true, includes each account id, balance, lock state, public key, and NFC UID (no secrets).',
  })
  @ApiResponse({ status: 200, description: 'User retrieved successfully', type: GetUserResponseDto })
  async getCurrentUser(
    @Query('includeAccounts', new DefaultValuePipe(false), ParseBoolPipe) includeAccounts: boolean,
    @Req() req: Request & { apiAuth?: ApiAuthContext },
  ): Promise<GetUserResponseDto> {
    const userId = req.apiAuth?.userId;
    if (!userId) {
      throw new UnauthorizedException('Missing auth context');
    }
    return await this.usersService.getUser(userId, { includeAccounts });
  }

  @UseInterceptors(new ResponseDtoOmitter(GetUserResponseDto))
  @Get(':id')
  @UseGuards(ApiTokenGuard)
  @RequirePrivilege(ApiTokenPrivilege.ADMIN)
  @ApiOperation({
    summary: 'Get a user',
    description: 'Get a user with the given ID',
  })
  @ApiParam({ name: 'id', description: 'User ID', format: 'uuid' })
  @ApiQuery({
    name: 'includeAccounts',
    required: false,
    type: Boolean,
    description: 'When true, includes each account id, balance, lock state, public key, and NFC UID (no secrets).',
  })
  @ApiResponse({ status: 200, description: 'User retrieved successfully', type: GetUserResponseDto })
  async getUser(
    @Param() params: GetUserRequestDto,
    @Query('includeAccounts', new DefaultValuePipe(false), ParseBoolPipe) includeAccounts: boolean,
  ): Promise<GetUserResponseDto> {
    const userId = params.id;
    return await this.usersService.getUser(userId, { includeAccounts });
  }
}